fix: separate useAvailability, fix multiple request issue
