+++ frontend/src/pages/profile/SchedulerControl.tsx
@@ -29,7 +29,8 @@ const SchedulerControl = ({
 
   // When values change, then we send a reuqest
   useEffect(() => {
-    const fromString = `${fromYear}-${fromMonth}-01`;
+    const fromString = dayjs(`${fromYear}/${fromMonth}/01`)
+      .format('YYYY-MM-DD');
     const toString = dayjs(`${toYear}/${toMonth}/01`)
       .endOf('month')
       .format('YYYY-MM-DD');
